Transfer SeqHist to a separate component
